# improving perf for mapping over ArrayObservable and ScalarObservables