For mapping over scalar observables:

was

> old map of scalar with immediate scheduler x 648,735 ops/sec ±0.54% (94 runs sampled)
> new map of scalar with immediate scheduler x 1,329,174 ops/sec ±1.32% (94 runs sampled)
>	104.89% faster than Rx2

now

> old map of scalar with immediate scheduler x 632,782 ops/sec ±0.69% (94 runs sampled)
> new map of scalar with immediate scheduler x 4,032,053 ops/sec ±0.52% (95 runs sampled)
>	537.19% faster than Rx2

For mapping over array observables:

was

> old map of array with immediate scheduler x 147,364 ops/sec ±0.79% (91 runs sampled)
> new map of array with immediate scheduler x 267,194 ops/sec ±0.65% (94 runs sampled)
> 	81.32% faster than Rx2

now

> old map of array with immediate scheduler x 143,522 ops/sec ±1.18% (94 runs sampled)
> new map of array with immediate scheduler x 1,063,957 ops/sec ±0.63% (96 runs sampled)
>	641.32% faster than Rx2