@blesh 2015-10-02
Okay, so the interesting thing here is you'll notice I had to change some tests...

One potential issue is that with these optimizations, if the user is doing something like mutating an outer-scoped variable in `map` and they're relying on that with something like `retry`, the mapping function really is only called once... In other words, an end user could be punished for using mutable state if I proceed with this PR.

For example:

```js
let arrObs = Observable.of(1,2,3,4,5,6);
let calls = 0;
let source = arrObs.map(x => {
   if(calls++ > 30) { throw 'bad' };
}).repeat();

source.subscribe();
```

In the example above, the mapping function will only really be called 6 times (one for each item in the original ArrayObservable returned by `of`), because `map` is returning an ArrayObservable directly, rather than a plain Observable that has been mapped.

I'd love input here as to whether or not this is a good idea @jhusain @mattpodwysocki @benjchristensen @headinthebox @steveorsomethin @trxcllnt