@robwormald 2015-10-02
It seems like an incredible performance boost. The 'gotcha' needs to be considered, but I'm in favor of faster things instead of slower things to cover use cases that are probably a bad idea anyway. 

My vote : add this and and document the case(s) where it might bite end-users. 537 percents is a lot of percents.