@blesh 2015-10-03
Okay, after discussions with @staltz and @jhusain, among others, that occurred offline, this looks good. Although @jhusain pointed out a really good idea, that I could probably just override map on the ScalarObservable rather than add checking here, so I'm going to do that, then merge.