@blesh 2015-10-07
One of the only reasons I haven't merged this yet is because I've got the heeby-jeebies about essentially removing the stream processing of Arrays (it becomes just Array.prototype.map under the hood, creating new array instances) I suspect there's a longer tail to this, were if you have `Observable.of(1,2,3)` and `map` on it 10 times or more consecutively, it ends up being faster to leave it as is. I think the *scalar* side of this is a huge win, but maybe not the array side...